backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
onnxruntime

tensorflow

# only needed for the one-time conversion (convert_to_onnx.py)
# tf2onnx
